Removed Accel and changed drift to demand
Removed the erroneous recording of floor accelerations (they were relative and not absolute) and change the output to be in terms of demand and not just drift. The Dc will need to be in radians from now on. A zero-length demand can be checked too and the Dc just becomes the threshold. Added more significant digits to the output file